rename types to filetypes
