+++ bundle/dicefile.go
@@ -25,7 +25,7 @@ import (
 	commonpb "github.com/erda-project/erda-proto-go/common/pb"
 	"github.com/erda-project/erda-proto-go/core/file/pb"
 	"github.com/erda-project/erda/bundle/apierrors"
-	"github.com/erda-project/erda/internal/core/file/types"
+	"github.com/erda-project/erda/internal/core/file/filetypes"
 	"github.com/erda-project/erda/pkg/http/httpclient"
 	"github.com/erda-project/erda/pkg/http/httputil"
 )
@@ -82,7 +82,7 @@ func (b *Bundle) DeleteDiceFile(uuid string) error {
 }
 
 // UploadFile 上传文件
-func (b *Bundle) UploadFile(req types.FileUploadRequest, clientTimeout ...int64) (*pb.File, error) {
+func (b *Bundle) UploadFile(req filetypes.FileUploadRequest, clientTimeout ...int64) (*pb.File, error) {
 	host, err := b.urls.ErdaServer()
 	if err != nil {
 		return nil, err
