+++ internal/apps/dop/providers/issue/core/file.go
@@ -40,6 +40,7 @@ import (
 	"github.com/erda-project/erda/internal/apps/dop/providers/issue/core/common"
 	"github.com/erda-project/erda/internal/apps/dop/providers/issue/dao"
 	"github.com/erda-project/erda/internal/apps/dop/services/apierrors"
+	"github.com/erda-project/erda/internal/core/file/filetypes"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/database/dbengine"
 	"github.com/erda-project/erda/pkg/excel"
@@ -265,7 +266,7 @@ func (i *IssueService) ExportExcelAsync(record *legacydao.TestFileRecord) {
 	}
 	f.Seek(0, 0)
 	expiredAt := time.Now().Add(time.Duration(conf.ExportIssueFileStoreDay()) * 24 * time.Hour)
-	uploadReq := types.FileUploadRequest{
+	uploadReq := filetypes.FileUploadRequest{
 		FileNameWithExt: fmt.Sprintf("%s.xlsx", tableName),
 		FileReader:      f,
 		From:            issueService,
@@ -591,7 +592,7 @@ func (i *IssueService) ExportExcel2(data [][]string, sheetName string) (string, 
 	if err := file.Write(&buff); err != nil {
 		return "", errors.Errorf("failed to write content, sheetName: %s, err: %v", sheetName, err)
 	}
-	diceFile, err := i.bdl.UploadFile(types.FileUploadRequest{
+	diceFile, err := i.bdl.UploadFile(filetypes.FileUploadRequest{
 		FileNameWithExt: sheetName + ".xlsx",
 		ByteSize:        int64(buff.Len()),
 		FileReader:      ioutil.NopCloser(&buff),
