+++ internal/apps/dop/services/project/project_package.go
@@ -45,6 +45,7 @@ import (
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/internal/apps/dop/services/apierrors"
 	"github.com/erda-project/erda/internal/apps/dop/services/namespace"
+	"github.com/erda-project/erda/internal/core/file/filetypes"
 	"github.com/erda-project/erda/pkg/filehelper"
 	"github.com/erda-project/erda/pkg/http/httputil"
 	"github.com/erda-project/erda/pkg/oauth2/tokenstore/mysqltokenstore"
@@ -254,7 +255,7 @@ func (t *PackageDataDirector) GenAndUploadZipPackage() (string, error) {
 	}
 	zipTmpFile.Seek(0, 0)
 	expiredAt := time.Now().Add(packageValidityPeriod)
-	uploadReq := types.FileUploadRequest{
+	uploadReq := filetypes.FileUploadRequest{
 		FileNameWithExt: t.Creator.GetPackageName(),
 		FileReader:      zipTmpFile,
 		From:            projectPackageResource,
@@ -599,7 +600,7 @@ func (t *PackageDataDirector) uploadArtifacZipFile(artifact *apistructs.Artifact
 	defer f.Close()
 
 	expiredAt := time.Now().Add(projectPackageValidityPeriod)
-	req := types.FileUploadRequest{
+	req := filetypes.FileUploadRequest{
 		FileReader:      f,
 		FileNameWithExt: artifact.ZipFile,
 		Creator:         t.Creator.GetContext().UserID,
@@ -688,7 +689,7 @@ func (p *Project) ImportPackage(req apistructs.ImportProjectPackageRequest, r *h
 		return 0, fmt.Errorf("project template file must be a zip package")
 	}
 	expiredAt := time.Now().Add(projectPackageValidityPeriod)
-	uploadReq := types.FileUploadRequest{
+	uploadReq := filetypes.FileUploadRequest{
 		FileNameWithExt: fileHeader.Filename,
 		FileReader:      f,
 		From:            projectPackageResource,
