+++ internal/tools/gittar/api/backup.go
@@ -19,7 +19,7 @@ import (
 	"os"
 
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/internal/core/file/types"
+	"github.com/erda-project/erda/internal/core/file/filetypes"
 	"github.com/erda-project/erda/internal/tools/gittar/helper"
 	"github.com/erda-project/erda/internal/tools/gittar/webcontext"
 )
@@ -57,7 +57,7 @@ func Backup(ctx *webcontext.Context) {
 		ctx.Abort(err)
 		return
 	}
-	req := types.FileUploadRequest{
+	req := filetypes.FileUploadRequest{
 		FileNameWithExt: branch + "." + format,
 		FileReader:      f,
 		From:            "backup",
